add bookview
