+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 // import * as BooksAPI from './BooksAPI'
 import './App.css'
-import ListBooks from './pages/ListBooks'
+import ListBooks from './components/ListBooks'
 
 class BooksApp extends React.Component {
   state = {
