+++ src/components/BookView.js
@@ -1,0 +1,36 @@
+import React, { Component } from 'react';
+
+
+class BookView extends Component {
+  // handleSubmit = (e) => {
+  //   e.preventDefault()
+  //   const values = serializeForm(e.target, { hash: true })
+  //   if (this.props.onCreateContact)
+  //     this.props.onCreateContact(values)
+  // }
+
+  render() {
+    const {title, authors, imageLinks} = this.props.book
+   
+    return (
+      <div className="book">
+      <div className="book-top">
+        <div className="book-cover" style={{ width: 128, height: 193, backgroundImage: `url(${imageLinks.smallThumbnail})` }}></div>
+        <div className="book-shelf-changer">
+          <select>
+            <option value="none" disabled>Move to...</option>
+            <option value="currentlyReading">Currently Reading</option>
+            <option value="wantToRead">Want to Read</option>
+            <option value="read">Read</option>
+            <option value="none">None</option>
+          </select>
+        </div>
+      </div>
+      <div className="book-title">{title}</div>
+      <div className="book-authors">{authors}</div>
+    </div>
+    )
+  }
+}
+
+export default BookView
